ingestor/requirements.txt
psycopg2-binary==2.9.9
openpyxl==3.1.2
requests==2.31.0
msal==1.26.0
apscheduler==3.10.4
python-dotenv==1.0.1
